backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pillow==10.4.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
google-generativeai>=0.8.3
pytest>=8.3.0
httpx>=0.27.0
